refactor(converter): reorganize tests
