+++ converter/tests/versions/create_health_case/test_v1_v2_converter.py
@@ -2,7 +2,7 @@ import json
 from unittest.mock import patch
 from converter.constants import Constants
 from converter.versions.create_case_health.create_case_health_converter import CreateHealthCaseConverter
-from .test_helpers import TestHelper, get_file_endpoint
+from tests.test_helpers import TestHelper, get_file_endpoint
 from snapshottest import TestCase
