+++ converter/tests/versions/test_resources_response_converter.py
@@ -1,53 +1,53 @@
 import json
 from snapshottest import TestCase
 from converter.constants import Constants
-from converter.versions.resources_status.resources_status_converter import ResourcesStatusConverter
-from .test_helpers import TestHelper, get_file_endpoint
+from converter.versions.resources_response.resources_response_converter import ResourcesResponseConverter
+from tests.test_helpers import TestHelper, get_file_endpoint
 
 def test_V1_to_V2_upgrade():
-    v2_schema_endpoint = get_file_endpoint(Constants.V2_GITHUB_TAG, Constants.RS_SR_TAG)
+    v2_schema_endpoint = get_file_endpoint(Constants.V2_GITHUB_TAG, Constants.RS_RR_TAG)
     v2_schema = TestHelper.load_json_file_online(v2_schema_endpoint)
 
     TestHelper.conversion_tests_runner(
-        sample_dir=Constants.RS_SR_TAG,
+        sample_dir=Constants.RS_RR_TAG,
         envelope_file=Constants.EDXL_HEALTH_TO_HEALTH_ENVELOPE_PATH,
-        converter_method=ResourcesStatusConverter.convert_v1_to_v2,
+        converter_method=ResourcesResponseConverter.convert_v1_to_v2,
         target_schema=v2_schema,
         online_tag=Constants.V1_GITHUB_TAG
     )
 
 def test_V2_to_V1_downgrade():
-    v1_schema_endpoint = get_file_endpoint(Constants.V1_GITHUB_TAG, Constants.RS_SR_TAG)
+    v1_schema_endpoint = get_file_endpoint(Constants.V1_GITHUB_TAG, Constants.RS_RR_TAG)
     v1_schema = TestHelper.load_json_file_online(v1_schema_endpoint)
 
     TestHelper.conversion_tests_runner(
-        sample_dir=Constants.RS_SR_TAG,
+        sample_dir=Constants.RS_RR_TAG,
         envelope_file=Constants.EDXL_HEALTH_TO_HEALTH_ENVELOPE_PATH,
-        converter_method=ResourcesStatusConverter.convert_v2_to_v1,
+        converter_method=ResourcesResponseConverter.convert_v2_to_v1,
         target_schema=v1_schema,
         online_tag=Constants.V2_GITHUB_TAG
     )
 
 def test_V2_to_V3_upgrade():
-    v3_schema_endpoint = get_file_endpoint(Constants.V3_GITHUB_TAG, Constants.RS_SR_TAG)
+    v3_schema_endpoint = get_file_endpoint(Constants.V3_GITHUB_TAG, Constants.RS_RR_TAG)
     v3_schema = TestHelper.load_json_file_online(v3_schema_endpoint)
 
     TestHelper.conversion_tests_runner(
-        sample_dir=Constants.RS_SR_TAG,
+        sample_dir=Constants.RS_RR_TAG,
         envelope_file=Constants.EDXL_HEALTH_TO_HEALTH_ENVELOPE_PATH,
-        converter_method=ResourcesStatusConverter.convert_v2_to_v3,
+        converter_method=ResourcesResponseConverter.convert_v2_to_v3,
         target_schema=v3_schema,
         online_tag=Constants.V2_GITHUB_TAG
     )
 
 def test_V3_to_V2_downgrade():
-    v2_schema_endpoint = get_file_endpoint(Constants.V2_GITHUB_TAG, Constants.RS_SR_TAG)
+    v2_schema_endpoint = get_file_endpoint(Constants.V2_GITHUB_TAG, Constants.RS_RR_TAG)
     v2_schema = TestHelper.load_json_file_online(v2_schema_endpoint)
 
     TestHelper.conversion_tests_runner(
-        sample_dir=Constants.RS_SR_TAG,
+        sample_dir=Constants.RS_RR_TAG,
         envelope_file=Constants.EDXL_HEALTH_TO_HEALTH_ENVELOPE_PATH,
-        converter_method=ResourcesStatusConverter.convert_v3_to_v2,
+        converter_method=ResourcesResponseConverter.convert_v3_to_v2,
         target_schema=v2_schema,
         online_tag=Constants.V3_GITHUB_TAG
     )
@@ -56,55 +56,39 @@ class TestSnapshotV1V2Converter(TestCase):
     def test_exhaustive_snapshot_V1_to_V2_upgrade(self):
         message = TestHelper.create_edxl_json_from_sample(
             Constants.EDXL_HEALTH_TO_HEALTH_ENVELOPE_PATH,
-            "tests/fixtures/RS-SR/RS-SR_V1.0_exhaustive_fill.json"
+            "tests/fixtures/RS-RR/RS-RR_V1.0_exhaustive_fill.json"
         )
-        output_data = ResourcesStatusConverter.convert_v1_to_v2(message)
+        output_data = ResourcesResponseConverter.convert_v1_to_v2(message)
         self.assertMatchSnapshot(json.dumps(output_data, indent=2))
 
     def test_required_fields_snapshot_V1_to_V2_upgrade(self):
         message = TestHelper.create_edxl_json_from_sample(
             Constants.EDXL_HEALTH_TO_HEALTH_ENVELOPE_PATH,
-            "tests/fixtures/RS-SR/RS-SR_V1.0_required_fields.json"
+            "tests/fixtures/RS-RR/RS-RR_V1.0_required_fields.json"
         )
-        output_data = ResourcesStatusConverter.convert_v1_to_v2(message)
+        output_data = ResourcesResponseConverter.convert_v1_to_v2(message)
         self.assertMatchSnapshot(json.dumps(output_data, indent=2))
 
     def test_exhaustive_snapshot_V2_to_V1_upgrade(self):
         message = TestHelper.create_edxl_json_from_sample(
             Constants.EDXL_HEALTH_TO_HEALTH_ENVELOPE_PATH,
-            "tests/fixtures/RS-SR/RS-SR_V2.0_exhaustive_fill.json"
+            "tests/fixtures/RS-RR/RS-RR_V2.0_exhaustive_fill.json"
         )
-        output_data = ResourcesStatusConverter.convert_v2_to_v1(message)
+        output_data = ResourcesResponseConverter.convert_v2_to_v1(message)
         self.assertMatchSnapshot(json.dumps(output_data, indent=2))
 
     def test_exhaustive_snapshot_V2_to_V3_upgrade(self):
         message = TestHelper.create_edxl_json_from_sample(
             Constants.EDXL_HEALTH_TO_HEALTH_ENVELOPE_PATH,
-            "tests/fixtures/RS-SR/RS-SR_V2.0_exhaustive_fill.json"
+            "tests/fixtures/RS-RR/RS-RR_V2.0_exhaustive_fill.json"
         )
-        output_data = ResourcesStatusConverter.convert_v2_to_v3(message)
+        output_data = ResourcesResponseConverter.convert_v2_to_v3(message)
         self.assertMatchSnapshot(json.dumps(output_data, indent=2))
 
     def test_required_fields_snapshot_V2_to_V3_upgrade(self):
         message = TestHelper.create_edxl_json_from_sample(
             Constants.EDXL_HEALTH_TO_HEALTH_ENVELOPE_PATH,
-            "tests/fixtures/RS-SR/RS-SR_V2.0_required_fields.json"
+            "tests/fixtures/RS-RR/RS-RR_V2.0_required_fields.json"
         )
-        output_data = ResourcesStatusConverter.convert_v2_to_v3(message)
-        self.assertMatchSnapshot(json.dumps(output_data, indent=2))
-
-    def test_exhaustive_snapshot_V3_to_V2_downgrade(self):
-        message = TestHelper.create_edxl_json_from_sample(
-            Constants.EDXL_HEALTH_TO_HEALTH_ENVELOPE_PATH,
-            "tests/fixtures/RS-SR/RS-SR_V3.0_exhaustive_fill.json"
-        )
-        output_data = ResourcesStatusConverter.convert_v3_to_v2(message)
-        self.assertMatchSnapshot(json.dumps(output_data, indent=2))
-
-    def test_required_fields_snapshot_V3_to_V2_downgrade(self):
-        message = TestHelper.create_edxl_json_from_sample(
-            Constants.EDXL_HEALTH_TO_HEALTH_ENVELOPE_PATH,
-            "tests/fixtures/RS-SR/RS-SR_V3.0_required_fields.json"
-        )
-        output_data = ResourcesStatusConverter.convert_v3_to_v2(message)
+        output_data = ResourcesResponseConverter.convert_v2_to_v3(message)
         self.assertMatchSnapshot(json.dumps(output_data, indent=2))
