+++ converter/tests/versions/test_rs_eda_maj_converter.py
@@ -1,6 +1,6 @@
 from converter.constants import Constants
 from converter.versions.create_case_health.create_case_health_update_converter import CreateHealthUpdateCaseConverter
-from .test_helpers import TestHelper, get_file_endpoint
+from tests.test_helpers import TestHelper, get_file_endpoint
 
 def test_V1_to_V2_upgrade():
     v2_schema_endpoint = get_file_endpoint(Constants.V2_GITHUB_TAG, Constants.RS_EDA_MAJ_TAG)
